requirements.txt
asgiref==3.4.1
attrs==21.4.0
boto3==1.26.8
botocore==1.29.8
CacheControl==0.12.11
cachetools==5.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
defusedxml==0.7.1
Django==4.0.1
django-admin-rangefilter==0.8.4
django-cors-headers==3.11.0
django-encrypted-field==1.0.4
django-enumchoicefield==3.0.1
django-environ==0.8.1
django-filter==22.1
django-phonenumber-field==7.0.0
django-regex-field==2.0.2
django-storages==1.13.1
django-templated-mail==1.1.1
django-user-agents==0.4.0
djangorestframework==3.13.1
djangorestframework-api-key==2.2.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
docusign-esign==3.17.0
drf-yasg==1.20.0
Faker==11.3.0
firebase-admin==6.0.1
flake8==4.0.1
google-api-core==2.11.0
google-api-python-client==2.71.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.7.3
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.1
grpcio==1.51.1
grpcio-status==1.51.1
gunicorn==19.7.1
httplib2==0.21.0
idna==3.3
importlib-metadata==6.0.0
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.0.3
jmespath==1.0.1
Markdown==3.4.1
MarkupSafe==2.0.1
mccabe==0.6.1
mock==4.0.3
msgpack==1.0.4
nose==1.3.7
oauthlib==3.2.0
packaging==21.3
passlib==1.7.4
phonenumbers==8.12.56
Pillow==9.0.1
pluggy==1.0.0
proto-plus==1.22.1
protobuf==4.21.12
psycopg2-binary==2.9.3
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pycryptodomex==3.15.0
pyflakes==2.4.0
PyJWT==2.6.0
pyparsing==3.0.6
pytest==7.1.0
pytest-django==4.5.2
python-dateutil==2.8.2
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
s3transfer==0.6.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
text-unidecode==1.3
tomli==2.0.1
ua-parser==0.15.0
uritemplate==4.1.1
urllib3==1.26.8
user-agents==2.2.0
zipp==3.11.0
